Refactoring for base structure with new tailwind styles modules
