+++ frontend/.gitignore
@@ -36,3 +36,5 @@ yarn-error.log*
 next-env.d.ts
 
 .vscode/*
+
+yarn.lock
