--- frontend/app/(main)/(components)/post-item/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PostItem'
