+++ frontend/components/Card/Card.module.scss
