+++ frontend/components/CardPlatform/CardPlatform.module.scss
