+++ frontend/components/Carousel/Carousel.module.scss
