+++ frontend/components/Footer/Footer.module.scss
