+++ frontend/components/Footer/index.ts
