+++ frontend/components/Section/Section.tsx
@@ -1,4 +1,4 @@
-import s from './styles.module.scss'
+import s from './Section.module.scss'
 import { FC, MutableRefObject } from 'react'
 import cn from 'classnames'
