--- frontend/features/articles/index.ts
@@ -1,2 +1,0 @@
-export type { default as INews } from './models/INews'
-export type { default as INewsCategory } from './models/INewsCategory'
