+++ frontend/features/games/components/FilterGames/index.ts
@@ -1,0 +1,1 @@
+export * from './FilterGames'
