+++ frontend/features/news/components/LatestNews/index.ts
@@ -1,0 +1,1 @@
+export * from './LatestNews'
