+++ frontend/features/news/components/NewsItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NewsItem'
