+++ frontend/features/news/components/NewsItem/styles.module.scss
