+++ frontend/features/news/models/INews.ts
