+++ frontend/features/news/models/INewsCategory.ts
