+++ frontend/features/news/models/index.ts
@@ -1,0 +1,2 @@
+export * from './INews'
+export * from './INewsCategory'
